protocol: consolidate options back into top-level epp package
