+++ options.go
@@ -1,9 +1,12 @@
 package epp
 
 import (
+	"context"
 	"crypto/tls"
+	"time"
 
 	"github.com/domainr/epp2/internal/config"
+	"github.com/domainr/epp2/schema"
 )
 
 // Options configure [TODO] with specific features.
@@ -22,4 +25,20 @@ func WithTLS(cfg *tls.Config) Options {
 
 func WithPipeline(depth int) Options {
 	return config.Pipeline(depth)
+}
+
+func WithContext(ctx context.Context) Options {
+	return config.Context{Context: ctx}
+}
+
+func WithKeepAlive(d time.Duration) Options {
+	return config.KeepAlive(d)
+}
+
+func WithTimeout(d time.Duration) Options {
+	return config.Timeout(d)
+}
+
+func WithSchema(schemas ...schema.Schema) Options {
+	return config.Schemas(schemas)
 }
